mon_app/requirements.txt
Flask==2.3.3
pywebview==4.4.1
Jinja2==3.1.2
Werkzeug==2.3.7
cryptography==41.0.7
bcrypt==4.1.2
pycryptodome==3.19.0
flask-session==0.5.0
